refactor: 중복 코드 함수화 및 컴포넌트화
