+++ next-converter/src/lib/requireAuth.ts
@@ -1,0 +1,10 @@
+import { auth } from '@/auth';
+import { redirect } from 'next/navigation';
+
+export default async function requireAuth() {
+  const session = await auth();
+  if (!session) {
+    redirect('/');
+  }
+  return session;
+}
